Move resource locator JS parts to Route bundle
