+++ packages/page/assets/js/containers/Form/fields/PageTreeRoute.js
@@ -4,7 +4,8 @@ import {observer} from 'mobx-react';
 import {action, observable} from 'mobx';
 import {userStore} from 'sulu-admin-bundle/stores';
 import {Grid} from 'sulu-admin-bundle/components';
-import {SingleSelection, ResourceLocator} from 'sulu-admin-bundle/containers';
+import {SingleSelection} from 'sulu-admin-bundle/containers';
+import {ResourceLocator} from 'sulu-route-bundle/containers';
 import {translate} from 'sulu-admin-bundle/utils';
 import type {IObservableValue} from 'mobx/lib/mobx';
 import type {FieldTypeProps} from 'sulu-admin-bundle/types';
