+++ packages/page/assets/js/package.json
@@ -18,6 +18,7 @@
         "@testing-library/react": "^12.1.5",
         "@testing-library/user-event": "^14.5.2",
         "sulu-admin-bundle": "file:../../../src/Sulu/Bundle/AdminBundle/Resources/js",
+        "sulu-route-bundle": "file:../../route/assets/js",
         "sulu-media-bundle": "file:../../../src/Sulu/Bundle/MediaBundle/Resources/js",
         "sulu-website-bundle": "file:../../../src/Sulu/Bundle/WebsiteBundle/Resources/js"
     },
@@ -27,6 +28,7 @@
         "react": "^16.2.0 || ^17.0.0",
         "sulu-admin-bundle": "*",
         "sulu-media-bundle": "*",
+        "sulu-route-bundle": "*",
         "sulu-website-bundle": "*"
     }
 }
