+++ packages/route/assets/js/components/ResourceLocator/index.js
