+++ packages/route/assets/js/components/ResourceLocator/resourceLocator.scss
@@ -1,4 +1,4 @@
-@import '../../containers/Application/colors.scss';
+@import 'sulu-admin-bundle/containers/Application/colors.scss';
 
 $textColor: $black;
