+++ packages/route/assets/js/components/ResourceLocator/tests/ResourceLocator.test.js
