+++ packages/route/assets/js/components/index.js
@@ -1,0 +1,6 @@
+// @flow
+import ResourceLocator from './ResourceLocator';
+
+export {
+    ResourceLocator,
+};
