+++ packages/route/assets/js/containers/Form/fields/ResourceLocator.js
@@ -4,12 +4,12 @@ import {action, comparer, computed, observable, reaction, toJS} from 'mobx';
 import {observer} from 'mobx-react';
 import ResourceLocatorComponent from '../../../components/ResourceLocator';
 import ResourceLocatorHistory from '../../../containers/ResourceLocatorHistory';
-import Requester from '../../../services/Requester';
-import {translate} from '../../../utils/Translator';
-import Button from '../../../components/Button';
-import userStore from '../../../stores/userStore';
+import Requester from 'sulu-admin-bundle/services/Requester';
+import {translate} from 'sulu-admin-bundle/utils/Translator';
+import Button from 'sulu-admin-bundle/components/Button';
+import userStore from 'sulu-admin-bundle/stores/userStore';
 import resourceLocatorStyles from './resourceLocator.scss';
-import type {FieldTypeProps} from '../../../types';
+import type {FieldTypeProps} from 'sulu-admin-bundle/types';
 
 const PART_TAG = 'sulu.rlp.part';
