+++ packages/route/assets/js/containers/Form/index.js
@@ -1,0 +1,6 @@
+// @flow
+import ResourceLocator from './fields/ResourceLocator';
+
+export {
+    ResourceLocator,
+};
