+++ packages/route/assets/js/containers/ResourceLocatorHistory/index.js
