+++ packages/route/assets/js/containers/ResourceLocatorHistory/resourceLocatorHistory.scss
@@ -1,4 +1,4 @@
-@import '../../containers/Application/variables.scss';
+@import 'sulu-admin-bundle/containers/Application/variables.scss';
 
 .resource-locator-history-overlay,
 .loader {
