+++ packages/route/assets/js/containers/index.js
@@ -1,0 +1,8 @@
+// @flow
+import { ResourceLocator } from './Form';
+import ResourceLocatorHistory from './ResourceLocatorHistory';
+
+export {
+    ResourceLocator,
+    ResourceLocatorHistory,
+};
